test(e2e): refactor to e2e_tests.rs file
